get basic character support working again.  These changes get accentchars.tex, oddchars.tex unicode.tex and overstrike.tex working again ... at least with Microsoft Word 2004 on a Mac.
git-svn-id: https://latex2rtf.svn.sourceforge.net/svnroot/latex2rtf/trunk@895 e562ef82-0334-0410-8083-dbe22b1e9a75

+++ cfg.h
@@ -27,7 +27,7 @@ typedef struct ConfigEntryT
 {
    const char  *TexCommand;
    const char  *RtfCommand;
-   size_t      original_id;
+   int      original_id;
 } ConfigEntryT;
 
 void ReadLanguage(char *lang);
